requirements.txt
langchain-core
langchain-community
langchain-google-genai
python-dotenv
pandas
sqlalchemy
langchain
google-generativeai==0.8.3
streamlit